Rename all utils to services
This matches the naming of the other components, communicating a little
more clearly that this are specific services, not general utilities.
