+++ code/part-two/processor/actions/cancel_offer.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const getAddress = require('../utils/addressing');
-const { decode, encode, reject } = require('../utils/helpers');
+const getAddress = require('../services/addressing');
+const { decode, encode, reject } = require('../services/helpers');
 
 
 const cancelOffer = (context, publicKey, { offer }) => {
