+++ code/part-two/processor/services/helpers.js
